chore(eval-output): move context_window utilities to shared core.utils
The context_window module (fit_to_context, fit_row_to_context, etc.)
contains generic text-fitting helpers that aren't specific to any one
evaluator. They lived under custom_prompt_evaluator/ for historical
reasons. Moves them to agentic_eval/core/utils/context_window.py so
peer evaluators can import from a neutral location instead of a
sibling evaluator's package.

CustomPromptEvaluator's two relative imports (``from .context_window``)
now point to the shared module. The old copy is deleted.

No behaviour change — file content moved verbatim, only the import
paths change.
